+++ imagedl/modules/sources/__init__.py
@@ -3,15 +3,18 @@ from .base import BaseImageClient
 from .bing import BingImageClient
 from .i360 import I360ImageClient
 from .nasa import NASAImageClient
+from .gbif import GBIFImageClient
 from .baidu import BaiduImageClient
 from .sogou import SogouImageClient
 from .yahoo import YahooImageClient
 from .weibo import WeiboImageClient
+from .yande import YandeImageClient
 from .pexels import PexelsImageClient
 from .google import GoogleImageClient
 from ..utils import BaseModuleBuilder
 from .yandex import YandexImageClient
 from .huaban import HuabanImageClient
+from .locgov import LocGovImageClient
 from .pixabay import PixabayImageClient
 from .dimtown import DimTownImageClient
 from .danbooru import DanbooruImageClient
@@ -41,6 +44,7 @@ class ImageClientBuilder(BaseModuleBuilder):
         'GelbooruImageClient': GelbooruImageClient,      'PexelsImageClient': PexelsImageClient,                      'PixabayImageClient': PixabayImageClient,           'DimTownImageClient': DimTownImageClient, 
         'StockSnapImageClient': StockSnapImageClient,    'FreeImagesImageClient': FreeImagesImageClient,              'LifeOfPixImageClient': LifeOfPixImageClient,       'GratisoGraphyImageClient': GratisoGraphyImageClient,
         'PicJumboImageClient': PicJumboImageClient,      'OpenverseImageClient': OpenverseImageClient,                'INaturalistImageClient': INaturalistImageClient,   'NASAImageClient': NASAImageClient,
+        'GBIFImageClient': GBIFImageClient,              'LocGovImageClient': LocGovImageClient,                      'YandeImageClient': YandeImageClient,
     }